move some files around for better structure, and also move some classes to thier own file and added functionality for dir config widget
